Add ability to include/change backend headers in event
Before it was not possible to change the header within the email
backend. Customized headers can be used to provide more information to
the system receiving the email message.

Now a configuration allows headers that can be set/overwritten by the event
body.

This includes an architectural change:
We define a new base class 'Backend' that every backend needs to
implement.
It specifies how sub classes have to be structured and the steps every
backend performs.