OTRS: Set default queue to 'Misc'
Before the default queue was "Queueebene1::Queueebene2" if not specified
in the config which is very unlikely to exist on any OTRS installation.
So if it had to be set correctly in the config or the header.

Now the default queue is 'Misc' which exists by default on any standard
OTRS installations